sparrow: insar display improvements
- color handling and general setup
- float, int -> float64, int64 (numpy v1.24)
- right display of NaN values
- adopt CPTHandler for Kite Scene display purposes
- update internal functions

+++ src/gui/sparrow/elements/active_faults.py
@@ -23,7 +23,7 @@ km = 1e3
 
 
 def color(x):
-    return num.array(plot.to01(plot.color(x)), dtype=num.float)
+    return num.array(plot.to01(plot.color(x)), dtype=num.float64)
 
 
 fault_color_themes = {
